pyproject.toml
[project]
name = "lulc-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "grad-cam>=1.5.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.8.1",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.1.78",
    "pandas>=2.1.3",
    "pillow>=10.1.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "scikit-image>=0.22.0",
    "scikit-learn>=1.3.2",
    "torch>=2.1.1",
    "torchvision>=0.16.1",
    "tqdm>=4.66.1",
    "ttach>=0.0.3",
    "uvicorn>=0.24.0",
]
